Level Manager, SubwayManager, and Load Out Systems complete
Create prefabs and systems for a Level Manager (loads you between subway and level), subway manager managers UI and exiting the subway, and WeaponCubes allow the player to swap between the three personas. exiting the subway starts a coroutine which fades to black, moves the player to the level starting position, starts a countdown, activates the enemy spawn systems, and then re-enables input to the player.